requirements.txt
matplotlib==3.9.2
notebook==7.2.2
pandas==2.2.3
plotly==5.24.1
pymongo==4.10.1
matplotlib==3.9.2
seaborn==0.13.2
black==24.10.0
python-dotenv==1.0.1
fastapi[standard]==0.115.0
python-binance
pytest-cov
pytest-asyncio
python-telegram-bot